pyproject.toml
[project]
name = "langgraph-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.16",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langsmith>=0.5.2",
    "fastmcp",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pyyaml>=6.0",
    "toml>=0.10.2",
]
